@@ -1,6 +1,8 @@
+// swift-tools-version: 6.0
 import PackageDescription
 
-protocol LibraryDescription {
-    static var library: Product { get }
+public protocol LibraryDescription {
+    static var product: Product { get }
     static var targets: [Target] { get }
+    static var testTargets: [Target] { get }
 }
