@@ -10,7 +10,9 @@ import GoogleSignInSwift
 // dependencies
 import AuthenticationClient
 import UserDatabaseClient
-import ErrorClient
+
+// features
+import ErrorFeature
 
 // models
 import UserModel
